backend/requirements/base.txt
# Django core
Django==5.0.9
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-debug-toolbar==4.4.6

# Database
psycopg[binary]==3.2.3

# Tree structure
django-mptt==0.16.0

# Filtering
django-filter==24.3

# CORS
django-cors-headers==4.6.0

# Redis & Celery
redis==5.2.0
celery==5.4.0
django-celery-beat==2.7.0

# HTTP client
httpx==0.28.1

# Validation
pydantic==2.10.3

# Production server
gunicorn==23.0.0
whitenoise==6.8.2

# Monitoring
django-prometheus==2.3.1

# Environment
python-decouple==3.8
